start regrouping scripts
